- 梳理代码

+++ Engine/Renderer/GLES3/src/GLES3CommandBindMesh.h
@@ -38,5 +38,5 @@ public:
 
 private:
 	CGfxMeshPtr m_ptrMesh;
-	uintptr_t m_indexDraw;
+	int m_indexDraw;
 };
